fstack: Fix display depth after [v]fork
Currently it assumes that the next rstack is the exit of [v]fork().  But
if kernel tracing was enabled, it'd see kernel functions first so the
stack count and display depth would be corrupted.

So keep the display depth at the time of fork() at parent, and use it in
a child.  Same goes to daemon(3) too which calls fork() internally.

However this didn't handle the user stack count so it would show
'negative stack count' debug message when return to user..

Signed-off-by: Namhyung Kim <namhyung@gmail.com>